move convertToReadableUnits to a helper function
